refactor: deprecate utility module
