+++ src/brisk/__init__.py
@@ -32,5 +32,4 @@ from brisk.training.workflow import Workflow
 from brisk.training.alert_mailer import AlertMailer
 from brisk.training.logging_util import TqdmLoggingHandler, FileFormatter
 from brisk.configuration.algorithm_wrapper import AlgorithmWrapper, AlgorithmCollection
-from brisk.utility.utility import create_metric
 from brisk.version import __version__
